Made message dispatching asynchronous
MessageDispatcher now has a pool of worker processes. The msg method
receives an additional argument called callback. When msg is called, the
appropriate plugins are executed in the pool. Then the plugin call
returns a value, the callback function is called with the return value
as the function's argument.

It's kind of a messy solution, but I'd rather have some workers
processing plugin calls instead of block the whole process.