Update message structure
MessageDispatcher doesn't need to know anything about the IRC message
content it receives. It uses a provided message filter for filtering
messages. The only filtering MessageDispatcher does is the regex pattern
matching.

When MessageDispatcher finds a matching message, it packs it into a
Message object along with the match dictionary and settings, and yields
the object with the handler.

BotCore packs each message it receives into a IrcMsg object. This object
becomes the content property of MessageDispatcher's Message object.
Updates message structure

MessageDispatcher doesn't need to know anything about the IRC message
content it receives. It uses a provided message filter for filtering
messages. The only filtering MessageDispatcher does is the regex pattern
matching.

When MessageDispatcher finds a matching message, it packs it into a
Message object along with the match dictionary and settings, and yields
the object with the handler.

BotCore packs each message it receives into a IrcMsg object. This object
becomes the content property of MessageDispatcher's Message object.